search-agent/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black>=25.9.0",
    "isort>=6.1.0",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.34",
    "langchain-tavily>=0.2.11",
    "python-dotenv>=1.1.1",
]

[tool.uv.workspace]
members = [
    "venv",
]
